Make AppServiceConsoleLogs use the new DefaultEventType
- Adds support for the LAQueryLogs, but no test yet
