+++ src/main/java/com/teragrep/nlf_01/NLFPlugin.java
@@ -109,8 +109,8 @@ public final class NLFPlugin implements Plugin {
             else if ("AppEvents".equals(type)) {
                 eventTypes.add(new AppEventsType(parsedEvent, realHostname, componentNameForPartitions));
             }
-            else if ("AppServiceConsoleLogs".equals(type)) {
-                eventTypes.add(new AppServiceConsoleLogsType(parsedEvent, realHostname, componentNameForPartitions));
+            else if (Set.of("AppServiceConsoleLogs", "LAQueryLogs").contains(type)) {
+                eventTypes.add(new DefaultEventType(parsedEvent, realHostname, componentNameForPartitions));
             }
             else if (Set.of("AppTraces", "AppDependencies", "AppExceptions", "AppRequests").contains(type)) {
                 eventTypes.add(new AppInsightType(parsedEvent, realHostname, componentNameForPartitions));
